refactor(site): decompose PlaygroundPage and remove dead code
PlaygroundPage shrinks 588 → 132 LOC by extracting six hooks
(usePlaygroundActions, usePlaygroundPanels, usePlaygroundCommands,
useEditorBridges, useShortcutHelpKey, useAutoExpandOnError) and a
DesktopLayout component, leaving an orchestrator that wires hooks
together. Behavior is preserved end-to-end (smoke test green).

Drops 319 LOC of orphaned landing-page code: useGitHubStars,
useInView, syntaxHighlight (none had any callers; App.tsx only
mounts PlaygroundPage). Tightens SHORTCUTS to a const-with-literal
keys, eliminating ~10 non-null assertions at callsites. Dedupes
the STL/STEP blob-download pattern in useCodeExecution into a
local downloadBlob helper.

Note: shiki and @types/lz-string remain in site/package.json as
unused deps. Removing them requires updating the root lockfile,
which npm < 11.11 (the local toolchain) silently strips top-level
@emnapi entries from. Deferring to a follow-up under npm 11.11+.
