--- site/src/hooks/useInView.ts
@@ -1,31 +1,0 @@
-import { useEffect, useRef, useState } from 'react';
-
-export function useInView<T extends HTMLElement = HTMLDivElement>(
-  options: { threshold?: number; once?: boolean } = {}
-): [React.RefObject<T | null>, boolean] {
-  const { threshold = 0.1, once = true } = options;
-  const ref = useRef<T | null>(null);
-  const [inView, setInView] = useState(false);
-
-  useEffect(() => {
-    const el = ref.current;
-    if (!el) return;
-
-    const observer = new IntersectionObserver(
-      ([entry]) => {
-        if (entry.isIntersecting) {
-          setInView(true);
-          if (once) observer.disconnect();
-        } else if (!once) {
-          setInView(false);
-        }
-      },
-      { threshold }
-    );
-
-    observer.observe(el);
-    return () => observer.disconnect();
-  }, [threshold, once]);
-
-  return [ref, inView];
-}
